@@ -4,10 +4,7 @@
     {
         public ApplicationUserDto(ApplicationUser applicationUser)
         {
-            if (applicationUser is null)
-            {
-                throw new ArgumentNullException(nameof(applicationUser));
-            }
+            ArgumentNullException.ThrowIfNull(applicationUser);
 
             this.Id = applicationUser.Id;
             this.Username = applicationUser.Username;
