@@ -11,10 +11,7 @@
     {
         public override InterceptionResult<int> SavingChanges(DbContextEventData eventData, InterceptionResult<int> result)
         {
-            if (eventData is null)
-            {
-                throw new ArgumentNullException(nameof(eventData));
-            }
+            ArgumentNullException.ThrowIfNull(eventData);
 
             TimestampInterceptor.SetAuditedTimestamps(eventData);
 
@@ -23,10 +20,7 @@
 
         public override ValueTask<InterceptionResult<int>> SavingChangesAsync(DbContextEventData eventData, InterceptionResult<int> result, CancellationToken cancellationToken = default)
         {
-            if (eventData is null)
-            {
-                throw new ArgumentNullException(nameof(eventData));
-            }
+            ArgumentNullException.ThrowIfNull(eventData);
 
             TimestampInterceptor.SetAuditedTimestamps(eventData);
 
