@@ -14,10 +14,7 @@ namespace EngineBay.Persistence
 
         public static new void CreateDataAnnotations(ModelBuilder modelBuilder)
         {
-            if (modelBuilder is null)
-            {
-                throw new ArgumentNullException(nameof(modelBuilder));
-            }
+            ArgumentNullException.ThrowIfNull(modelBuilder);
 
             modelBuilder.Entity<ApplicationUser>().ToTable(typeof(ApplicationUser).Name.Pluralize());
 
